+++ R/aemet_tools.R
@@ -1,23 +1,39 @@
 ################################################################################################################
 # Author: Manuel Pizarro <m.pizarro@csic.es>
 # Ecosystem Conservation, IPE (CSIC) <http://www.ipe.csic.es/conservacion-bio/>
-# Version: 0.1.0
+# Version: 0.2.0
 ################################################################################################################
 
 #' @title Last observation values for a station
-#' @description Get last observation values for a station
 #'
-#' @param station Station identifier code (see \code{\link{aemet_stations}}).
-#' @param apikey Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
+#' @description Get last observation values for a station.
 #'
-#' @return a data.frame
-#' @export
+#' @param station Character string as station identifier code (see \code{\link{aemet_stations}}).
+#' @param apikey Character string as personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
+#'
+#' @return a data.frame.
+#'
+#' @importFrom methods missingArg
 #'
 #' @examples \dontrun{
 #' aemet_last_obs("9434", apikey)
 #' }
+#'
+#' @export
 
 aemet_last_obs <- function(station, apikey) {
+
+  if (missingArg(station))
+    stop("Station can't be missing")
+
+  if (missingArg(apikey))
+    stop("API key can't be missing")
+
+  if (!is.character(station))
+    stop("Station need to be character string")
+
+  if (!is.character(apikey))
+    stop("API key need to be character string")
 
   apidest <- paste0("/api/observacion/convencional/datos/estacion/", station)
 
@@ -26,21 +42,37 @@ aemet_last_obs <- function(station, apikey) {
 }
 
 #' @title Normal climatology values for a station
+#'
 #' @description Get normal climatology values for a station. Standard climatology from 1981 to 2010.
 #'
 #' @note Code modified from project https://github.com/SevillaR/aemet
 #'
-#' @param station Station identifier code (see \code{\link{aemet_stations}}).
-#' @param apikey Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
+#' @param station Character string as station identifier code (see \code{\link{aemet_stations}}).
+#' @param apikey Character string as personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
 #'
-#' @return a data.frame
-#' @export
+#' @return a data.frame.
+#'
+#' @importFrom methods missingArg
 #'
 #' @examples \dontrun{
 #' aemet_normal_clim("9434", apikey)
 #' }
+#'
+#' @export
 
 aemet_normal_clim <- function(station, apikey) {
+
+  if (missingArg(station))
+    stop("Station can't be missing")
+
+  if (missingArg(apikey))
+    stop("API key can't be missing")
+
+  if (!is.character(station))
+    stop("Station need to be character string")
+
+  if (!is.character(apikey))
+    stop("API key need to be character string")
 
   apidest <- paste0("/api/valores/climatologicos/normales/estacion/", station)
 
@@ -48,22 +80,50 @@ aemet_normal_clim <- function(station, apikey) {
 
 }
 
-#' @title Daily climatology values for a station
-#' @description Get daily climatology values for a station
+#' @title Daily/annual climatology values for a station
 #'
-#' @param station Station identifier code (see \code{\link{aemet_stations}}).
-#' @param apikey Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
-#' @param start a start date (format: %Y%m%d)
-#' @param end a end date (format: %Y%m%d)
+#' @description Get daily climatology values for a station for a maximum period of one year.
 #'
-#' @return a data.frame
-#' @export
+#' @param station Character string as station identifier code (see \code{\link{aemet_stations}}).
+#' @param apikey Character string as personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
+#' @param start Character string as start date (format: %Y%m%d).
+#' @param end Character string as end date (format: %Y%m%d).
+#'
+#' @return a data.frame.
+#'
+#' @importFrom methods missingArg
 #'
 #' @examples \dontrun{
-#' aemet_daily_clim("9434", apikey, 2000-01-01, 2000-12-31)
+#' aemet_daily_clim("9434", apikey, "2000-01-01", "2000-12-31")
 #' }
+#'
+#' @export
 
 aemet_daily_clim <- function(station, apikey, start, end) {
+
+  if (missingArg(station))
+    stop("Station can't be missing")
+
+  if (missingArg(apikey))
+    stop("API key can't be missing")
+
+  if (missingArg(start))
+    stop("Start year can't be missing")
+
+  if (missingArg(end))
+    stop("End year can't be missing")
+
+  if (!is.character(station))
+    stop("Station need to be character string")
+
+  if (!is.character(apikey))
+    stop("API key need to be character string")
+
+  if (!is.character(start))
+    stop("Start year need to be character string")
+
+  if (!is.character(end))
+    stop("End year need to be character string")
 
   apidest <- paste0("/api/valores/climatologicos/diarios/datos/fechaini/", start, "T00:00:00UTC/fechafin/", end, "T23:59:59UTC/estacion/", station)
 
@@ -72,20 +132,42 @@ aemet_daily_clim <- function(station, apikey, start, end) {
 }
 
 #' @title Monthly/annual climatology values for a station
-#' @description Get monthly/annual climatology values for a station
 #'
-#' @param station Station identifier code (see \code{\link{aemet_stations}}).
-#' @param apikey Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
-#' @param year a date (format: %Y)
+#' @description Get monthly/annual climatology values for a station.
 #'
-#' @return a data.frame
-#' @export
+#' @param station Character string as station identifier code (see \code{\link{aemet_stations}}).
+#' @param apikey Character string as personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
+#' @param year Numeric value as date (format: %Y).
+#'
+#' @return a data.frame.
+#'
+#' @importFrom methods missingArg
 #'
 #' @examples \dontrun{
 #' aemet_monthly_clim("9434", apikey, 2000)
 #' }
+#'
+#' @export
 
 aemet_monthly_clim <- function(station, apikey, year) {
+
+  if (missingArg(station))
+    stop("Station can't be missing")
+
+  if (missingArg(apikey))
+    stop("API key can't be missing")
+
+  if (missingArg(year))
+    stop("Year can't be missing")
+
+  if (!is.character(station))
+    stop("Station need to be character string")
+
+  if (!is.character(apikey))
+    stop("API key need to be character string")
+
+  if (!is.numeric(year))
+    stop("Year need to be numeric")
 
   apidest <- paste0("/api/valores/climatologicos/mensualesanuales/datos/anioini/", year, "/aniofin/", year, "/estacion/", station)
 
@@ -94,20 +176,42 @@ aemet_monthly_clim <- function(station, apikey, year) {
 }
 
 #' @title Extreme values for a station
-#' @description Get recorded extreme values for a station
 #'
-#' @param station Station identifier code (see \code{\link{aemet_stations}}).
-#' @param apikey Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
-#' @param parameter Temperature (T), Precipitatation (P) or Wind (W)
+#' @description Get recorded extreme values for a station.
 #'
-#' @return a data.frame
-#' @export
+#' @param station Character string as station identifier code (see \code{\link{aemet_stations}}).
+#' @param apikey Character string as personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
+#' @param parameter Character string as temperature (T), precipitatation (P) or wind (W) parameter.
+#'
+#' @return a data.frame.
+#'
+#' @importFrom methods missingArg
 #'
 #' @examples \dontrun{
 #' aemet_extremes_clim("9434", apikey, "T")
 #' }
+#'
+#' @export
 
 aemet_extremes_clim <- function(station, apikey, parameter = c("T", "P", "V")) {
+
+  if (missingArg(station))
+    stop("Station can't be missing")
+
+  if (missingArg(apikey))
+    stop("API key can't be missing")
+
+  if (missingArg(parameter))
+    stop("Parameter can't be missing")
+
+  if (!is.character(station))
+    stop("Station need to be character string")
+
+  if (!is.character(apikey))
+    stop("API key need to be character string")
+
+  if (!is.character(parameter))
+    stop("Parameter need to be character string")
 
   apidest <- paste0("/api/valores/climatologicos/valoresextremos/parametro/", parameter, "/estacion/", station)
 
@@ -116,20 +220,30 @@ aemet_extremes_clim <- function(station, apikey, parameter = c("T", "P", "V")) {
 }
 
 #' @title AEMET stations
-#' @description Get AEMET stations
+#'
+#' @description Get AEMET stations.
 #'
 #' @note Code modified from project https://github.com/SevillaR/aemet
 #'
-#' @param apikey String AEMET API key (see https://opendata.aemet.es/centrodedescargas/obtencionAPIKey).
+#' @param apikey Character string as API key (https://opendata.aemet.es/centrodedescargas/obtencionAPIKey).
 #'
-#' @return A data.frame
-#' @export
+#' @return a data.frame.
+#'
+#' @importFrom methods missingArg
 #'
 #' @examples \dontrun{
 #' stations <- aemet_stations(apikey)
 #' }
+#'
+#' @export
 
 aemet_stations <- function(apikey) {
+
+  if (missingArg(apikey))
+    stop("API key can't be missing")
+
+  if (!is.character(apikey))
+    stop("API key need to be character string")
 
   stations <- get_data_aemet(apidest = "/api/valores/climatologicos/inventarioestaciones/todasestaciones", apikey)
 
@@ -144,20 +258,24 @@ aemet_stations <- function(apikey) {
 }
 
 #' @title Client tool for AEMET API
-#' @description Client tool to get data from AEMET and convert json to data.frame
+#'
+#' @description Client tool to get data from AEMET and convert json to data.frame.
 #'
 #' @note Code modified from project https://github.com/vegmod/meteoland
 #'
-#' @param apidest character Destination URL. See \url{https://opendata.aemet.es/dist/index.html}.
-#' @param apikey Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
+#' @param apidest Character string as destination URL. See \url{https://opendata.aemet.es/dist/index.html}.
+#' @param apikey Character string as personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
 #' @param verbose True/False. Provides information about the flow of
 #' information between the client and server
 #'
-#' @return a data.frame
-#' @export
+#' @return a data.frame.
 #'
 #' @import httr
 #' @importFrom jsonlite fromJSON
+#' @importFrom methods missingArg
+#'
+#' @export
+
 
 get_data_aemet <- function(apidest, apikey, verbose = FALSE) {
 
@@ -216,21 +334,31 @@ get_data_aemet <- function(apidest, apikey, verbose = FALSE) {
   }
 }
 
-#' @title Converts ddmmss to decimal degrees
-#' @description Converts degrees, minutes and seconds to decimal degrees
+#' @title Converts dms to decimal degrees
+#'
+#' @description Converts degrees, minutes and seconds to decimal degrees.
 #'
 #' @note Code modified from project https://github.com/SevillaR/aemet
 #'
-#' @param input character
+#' @param input Character string as DMS coordinates.
 #'
-#' @return num
-#' @export
+#' @return a numeric value.
+#'
+#' @importFrom methods missingArg
 #'
 #' @examples \donttest{
 #' dms2decdegrees("055245W")
 #' }
+#'
+#' @export
 
 dms2decdegrees <- function(input) {
+
+  if (missingArg(input))
+    stop("Input can't be missing")
+
+  if (!is.character(input))
+    stop("Input need to be character string")
 
   deg <- as.numeric(substr(input, 0, 2))
   min <- as.numeric(substr(input, 3,4))
@@ -243,20 +371,30 @@ dms2decdegrees <- function(input) {
 }
 
 #' @title Normal climatology values for all stations
-#' @description Get normal climatology values for all stations
 #'
-#' @param apikey Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
+#' @description Get normal climatology values for all stations.
 #'
-#' @return a data.frame
-#' @export
+#' @param apikey Character string as personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
+#'
+#' @return a data.frame.
+#'
+#' @importFrom dplyr bind_rows
+#' @importFrom methods missingArg
 #'
 #' @examples \dontrun{
 #' aemet_normal_clim_all(apikey)
 #' }
 #'
-#' @importFrom dplyr bind_rows
+#' @export
+
 
 aemet_normal_clim_all <- function(apikey) {
+
+  if (missingArg(apikey))
+    stop("API key can't be missing")
+
+  if (!is.character(apikey))
+    stop("API key need to be character string")
 
   stations <- aemet_stations(apikey)
 
@@ -278,24 +416,52 @@ aemet_normal_clim_all <- function(apikey) {
 
 }
 
-#' @title Monthly climatology values for a station
-#' @description Get monthly climatology values for a period of years for a station
+#' @title Monthly climatology values of a station for a time period
 #'
-#' @param station Station identifier code (see \code{\link{aemet_stations}}).
-#' @param apikey Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
-#' @param start a start year (format: %Y)
-#' @param end a end year (format: %Y)
+#' @description Get monthly climatology values for a period of years for a station.
 #'
-#' @return a data.frame
-#' @export
+#' @param station Character string as station identifier code (see \code{\link{aemet_stations}}).
+#' @param apikey Character string as personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
+#' @param start Numeric value as start year (format: %Y).
+#' @param end a Numeric value as end year (format: %Y).
+#'
+#' @return a data.frame.
+#'
+#' @importFrom dplyr bind_rows
+#' @importFrom methods missingArg
 #'
 #' @examples \dontrun{
 #' aemet_monthly_period("9434", apikey, 2000, 2010)
 #' }
 #'
-#' @importFrom dplyr bind_rows
+#' @export
+
 
 aemet_monthly_period <- function(station, apikey, start, end) {
+
+  if (missingArg(station))
+    stop("Station can't be missing")
+
+  if (missingArg(apikey))
+    stop("API key can't be missing")
+
+  if (missingArg(start))
+    stop("Start year can't be missing")
+
+  if (missingArg(end))
+    stop("End year can't be missing")
+
+  if (!is.character(station))
+    stop("Station need to be character string")
+
+  if (!is.character(apikey))
+    stop("API key need to be character string")
+
+  if (!is.numeric(start))
+    stop("Start year need to be numeric")
+
+  if (!is.numeric(end))
+    stop("End year need to be numeric")
 
   data_all = data.frame()
 
@@ -316,23 +482,45 @@ aemet_monthly_period <- function(station, apikey, start, end) {
 
 }
 
-#' @title Monthly climatology for all stations
-#' @description Get monthly climatology values for a period of years for all stations
+#' @title Monthly climatology of all stations for a period of time
 #'
-#' @param apikey Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
-#' @param start a start year (format: %Y)
-#' @param end a end year (format: %Y)
+#' @description Get monthly climatology values for a period of years for all stations.
 #'
-#' @return a data.frame
-#' @export
+#' @param apikey Character string as personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
+#' @param start Numeric value as start year (format: %Y).
+#' @param end Numeric value as end year (format: %Y).
+#'
+#' @return a data.frame.
+#'
+#' @importFrom dplyr bind_rows
+#' @importFrom methods missingArg
 #'
 #' @examples \dontrun{
 #' aemet_monthly_period_all(apikey, 2000, 2010)
 #' }
 #'
-#' @importFrom dplyr bind_rows
+#' @export
+
 
 aemet_monthly_period_all <- function(apikey, start, end) {
+
+  if (missingArg(apikey))
+    stop("API key can't be missing")
+
+  if (missingArg(start))
+    stop("Start year can't be missing")
+
+  if (missingArg(end))
+    stop("End year can't be missing")
+
+  if (!is.character(apikey))
+    stop("API key need to be character string")
+
+  if (!is.numeric(start))
+    stop("Start year need to be numeric")
+
+  if (!is.numeric(end))
+    stop("End year need to be numeric")
 
   stations <- aemet_stations(apikey)
 
@@ -355,79 +543,134 @@ aemet_monthly_period_all <- function(apikey, start, end) {
 }
 
 ## R CMD check: the .'s that appear in pipelines
+
 if(getRversion() >= "2.15.1")  utils::globalVariables(c("."))
 
 #' @title First day of year
-#' @description Get first day of year
 #'
-#' @param year a start year (format: %Y)
-#' @param month a start year (format: %m)
+#' @description Get first day of year.
 #'
-#' @return date (format: %Y%m%d)
-#' @export
+#' @param year Numeric value as year (format: %Y).
+#'
+#' @return Character string as date (format: %Y%m%d).
+#'
+#' @import dplyr
+#' @import lubridate
+#' @importFrom rlang .data
+#' @importFrom methods missingArg
 #'
 #' @examples \donttest{
 #' first_day_of_year(2000)
 #' }
 #'
-#' @import dplyr
-#' @import lubridate
-#' @importFrom rlang .data
+#' @export
 
-first_day_of_year <- function(year, month = 1){
+
+first_day_of_year <- function(year){
+
+  if (missingArg(year))
+    stop("Year can't be missing")
+
+  if (!is.numeric(year))
+    stop("Year need to be numeric")
+
+  month = 1
 
   date <- ymd(paste0(year, "-", month, "-1")) %>%
   ceiling_date(., "month") %>% {.-days(31)}
+
+  date <- as.character(date)
 
   return(date)
 
 }
 
 #' @title Last day of year
-#' @description Get last day of year
 #'
-#' @param year a start year (format: %Y)
-#' @param month a start year (format: %m)
+#' @description Get last day of year.
 #'
-#' @return date (format: %Y%m%d)
-#' @export
+#' @param year Numeric value as year (format: %Y).
+#'
+#' @return Character string as date (format: %Y%m%d).
+#'
+#' @import dplyr
+#' @import lubridate
+#' @importFrom rlang .data
+#' @importFrom methods missingArg
 #'
 #' @examples \donttest{
 #' last_day_of_year(2000)
 #' }
 #'
-#' @import dplyr
-#' @import lubridate
-#' @importFrom rlang .data
+#' @export
 
-last_day_of_year <- function(year, month = 12){
+
+last_day_of_year <- function(year){
+
+  if (missingArg(year))
+    stop("Year can't be missing")
+
+  if (!is.numeric(year))
+    stop("Year need to be numeric")
+
+  month = 12
 
   date <- ymd(paste0(year, "-", month, "-1")) %>%
   ceiling_date(., "month") %>% {.-days(1)}
+
+  date <- as.character(date)
 
   return(date)
 
 }
 
-#' @title Daily climatology values for a station
-#' @description Get daily climatology values for a period of years for a station
+#' @title Daily climatology values of a station for a time period
 #'
-#' @param station Station identifier code (see \code{\link{aemet_stations}}).
-#' @param apikey Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
-#' @param start a start year (format: %Y)
-#' @param end a end year (format: %Y)
+#' @description Get daily climatology values for a period of years for a station.
 #'
-#' @return a data.frame
-#' @export
+#' @param station Character string as station identifier code (see \code{\link{aemet_stations}}).
+#' @param apikey Character string as personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
+#' @param start Numeric value as start year (format: %Y).
+#' @param end Numeric value as end year (format: %Y).
+#'
+#' @return a data.frame.
+#'
+#' @import lubridate
+#' @importFrom dplyr bind_rows
+#' @importFrom methods missingArg
 #'
 #' @examples \dontrun{
 #' aemet_daily_period("9434", apikey, 2000, 2010)
 #' }
 #'
-#' @import lubridate
-#' @importFrom dplyr bind_rows
+#' @export
+
 
 aemet_daily_period <- function(station, apikey, start, end) {
+
+  if (missingArg(station))
+    stop("Station can't be missing")
+
+  if (missingArg(apikey))
+    stop("API key can't be missing")
+
+  if (missingArg(start))
+    stop("Start year can't be missing")
+
+  if (missingArg(end))
+    stop("End year can't be missing")
+
+  if (!is.character(station))
+    stop("Station need to be character string")
+
+  if (!is.character(apikey))
+    stop("API key need to be character string")
+
+  if (!is.numeric(start))
+    stop("Start year need to be numeric")
+
+  if (!is.numeric(end))
+    stop("End year need to be numeric")
 
   data_all = data.frame()
 
@@ -448,24 +691,45 @@ aemet_daily_period <- function(station, apikey, start, end) {
 
 }
 
-#' @title Daily climatology values for all stations
-#' @description Get daily climatology values for a period of years for all stations
+#' @title Daily climatology values of all stations for a time period
 #'
-#' @param apikey Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
-#' @param start a start year (format: %Y)
-#' @param end a end year (format: %Y)
+#' @description Get daily climatology values for a period of years for all stations.
 #'
-#' @return a data.frame
-#' @export
+#' @param apikey Character string as personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).
+#' @param start Numeric value as start year (format: %Y).
+#' @param end Numeric value as end year (format: %Y).
+#'
+#' @return a data.frame.
+#'
+#' @import lubridate
+#' @importFrom dplyr bind_rows
+#' @importFrom methods missingArg
 #'
 #' @examples \dontrun{
 #' aemet_daily_period_all(apikey, 2000, 2010)
 #' }
 #'
-#' @import lubridate
-#' @importFrom dplyr bind_rows
+#' @export
 
 aemet_daily_period_all <- function(apikey, start, end) {
+
+  if (missingArg(apikey))
+    stop("API key can't be missing")
+
+  if (missingArg(start))
+    stop("Start year can't be missing")
+
+  if (missingArg(end))
+    stop("End year can't be missing")
+
+  if (!is.character(apikey))
+    stop("API key need to be character string")
+
+  if (!is.numeric(start))
+    stop("Start year need to be numeric")
+
+  if (!is.numeric(end))
+    stop("End year need to be numeric")
 
   stations <- aemet_stations(apikey)