+++ NEWS.md
@@ -1,3 +1,17 @@
+# climaemet 0.2.0
+
+* rename ggstripes_station() to climatestripes_station().
+
+* climatogram_normal(). New function to plot a Walter & Lieth climatic diagram from normal climatology values.
+
+* climatogram_period(). New function to plot a Walter & Lieth climatic diagram for a specified time period.
+
+* ggwindrose(). New function to plot windrose diagram.
+
+* windrose_days(). New function to plot a windrose (speed/direction) diagram of a station over a days period.
+
+* windrose_period(). New function to plot a windrose (speed/direction) diagram of a station over a time period.
+
 # climaemet 0.1.0
 
-* Added a `NEWS.md` file to track changes to the package.
+* Firts release July 2020.